+++ packages/utils/src/hex.ts
@@ -1,4 +1,5 @@
 import { Buffer } from 'buffer';
+import { encode } from 'cbor-x';
 
 export function decodeHex(s: string): Uint8Array {
   return Uint8Array.from(Buffer.from(s, 'hex'));
@@ -7,3 +8,10 @@ export function decodeHex(s: string): Uint8Array {
 export function encodeHex(arr: Uint8Array): string {
   return Buffer.from(arr).toString('hex');
 }
+
+/**
+ * convert a hex to cbor
+ * @param hex
+ */
+export const hexToCbor = (hex: string) =>
+  Buffer.from(encode(Buffer.from(hex, 'hex'))).toString('hex');